+++ package.json
@@ -18,12 +18,18 @@
     "html-webpack-plugin": "^5.5.1",
     "webpack": "^5.79.0",
     "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.13.3"
+    "webpack-dev-server": "^4.13.3",
+    "css-loader": "6.7.3",
+    "style-loader": "3.3.2"
   },
   "engines": {
     "node": "18"
   },
-  "browserslist": "defaults",
+  "browserslist": ["last 2 versions",
+  "not dead",
+  "not ie <= 11",
+  "not op_mini all",
+  "not < 0.25%"],
   "dependencies": {
     "dayjs": "1.11.7",
     "flatpickr": "^4.6.13"
