+++ src/view/edit-form-view.js
@@ -1,0 +1,177 @@
+import AbstractView from '../framework/view/abstract-view.js';
+import { destinations } from '../data/static-data.js';
+import dayjs from 'dayjs';
+import { offersByTypes } from '../data/static-data.js';
+
+function editFormTemplate(point) {
+  return `<li class="trip-events__item">
+  <form class="event event--edit" action="#" method="post">
+    <header class="event__header">
+      <div class="event__type-wrapper">
+        <label class="event__type  event__type-btn" for="event-type-toggle-1">
+          <span class="visually-hidden">Choose event type</span>
+          <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
+        </label>
+        <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
+
+        <div class="event__type-list">
+          <fieldset class="event__type-group">
+            <legend class="visually-hidden">Event type</legend>
+
+            <div class="event__type-item">
+              <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi" ${point.type === 'taxi' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-bus-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus" ${point.type === 'bus' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--bus" for="event-type-bus-1">Bus</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-train-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train" ${point.type === 'train' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--train" for="event-type-train-1">Train</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-ship-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship" ${point.type === 'ship' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--ship" for="event-type-ship-1">Ship</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-drive-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive" ${point.type === 'drive' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--drive" for="event-type-drive-1">Drive</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" ${point.type === 'flight' ? 'checked' : ''}>
+              <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
+              <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-sightseeing-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
+              <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-1">Sightseeing</label>
+            </div>
+
+            <div class="event__type-item">
+              <input id="event-type-restaurant-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
+              <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-1">Restaurant</label>
+            </div>
+          </fieldset>
+        </div>
+      </div>
+
+      <div class="event__field-group  event__field-group--destination">
+        <label class="event__label  event__type-output" for="event-destination-1">
+          Flight
+        </label>
+        <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Chamonix" list="destination-list-1">
+        <datalist id="destination-list-1">
+          ${destinations.map((dest) => dest.name).map((destName) => `<option value="${destName}"></option>`).join('')}
+        </datalist>
+      </div>
+
+      <div class="event__field-group  event__field-group--time">
+        <label class="visually-hidden" for="event-start-time-1">From</label>
+        <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="18/03/19 12:25">
+        &mdash;
+        <label class="visually-hidden" for="event-end-time-1">To</label>
+        <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="18/03/19 13:35">
+      </div>
+
+      <div class="event__field-group  event__field-group--price">
+        <label class="event__label" for="event-price-1">
+          <span class="visually-hidden">Price</span>
+          &euro;
+        </label>
+        <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="160">
+      </div>
+
+      <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
+      <button class="event__reset-btn" type="reset">Delete</button>
+      <button class="event__rollup-btn" type="button">
+        <span class="visually-hidden">Close event</span>
+      </button>
+    </header>
+    <section class="event__details">
+      <section class="event__section  event__section--offers">
+        <h3 class="event__section-title  event__section-title--offers">Offers</h3>
+
+        <div class="event__available-offers"></div>
+      </section>
+
+      <section class="event__section  event__section--destination">
+        <h3 class="event__section-title  event__section-title--destination">Destination</h3>
+        <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
+      </section>
+    </section>
+  </form>
+</li>`;
+}
+
+export default class EditFormView extends AbstractView {
+
+  constructor({point}, handleSave, handleCloseBtn) {
+    super();
+    this.point = point;
+
+    this.element.querySelector('.event__type-icon').src = `img/icons/${point.type}.png`;
+    this.element.querySelector('.event__type-output').textContent = point.type;
+    const destination = destinations.find((dest) => dest.id === point.destination);
+    this.element.querySelector('#event-destination-1').value = destination.name;
+    this.element.querySelector('#event-start-time-1').value = dayjs(point.dateFrom).format('DD/MM/YY HH:mm');
+    this.element.querySelector('#event-end-time-1').value = dayjs(point.dateTo).format('DD/MM/YY HH:mm');
+    this.element.querySelector('#event-price-1').value = point.basePrice;
+    this.element.querySelector('.event__destination-description').textContent = destination.description;
+    const offers = offersByTypes.find((offer) => offer.type === point.type).offers;
+
+    const offersContainer = this.element.querySelector('.event__available-offers');
+    for (const offer of offers) {
+      const offerContainer = document.createElement('div');
+      offerContainer.classList.add('event__offer-selector');
+
+      offerContainer.innerHTML = `
+        <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offer.id}" type="checkbox" name="${offer.id}" ${point.offers.includes(offer.id) ? 'checked' : ''}>
+        <label class="event__offer-label" for="event-offer-${offer.id}">
+          <span class="event__offer-title">${offer.title}</span>
+          &plus;&euro;&nbsp;
+          <span class="event__offer-price">${offer.price}</span>
+        </label>
+      `;
+      offersContainer.appendChild(offerContainer);
+    }
+
+
+    const form = this.element.querySelector('form');
+    form.addEventListener('change', () => {
+      const formData = new FormData(form);
+      const eventType = formData.get('event-type');
+      this.element.querySelector('.event__type-icon').src = `img/icons/${eventType}.png`;
+      this.element.querySelector('.event__type-output').textContent = eventType;
+    });
+
+    form.addEventListener('submit', (evt) => {
+      evt.preventDefault();
+      handleSave(new FormData(form));
+    });
+
+    form.addEventListener('keyup', (evt) => {
+      if (evt.key === 'Escape') {
+        handleCloseBtn(this);
+      }
+    });
+
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', () => {
+      handleCloseBtn(this);
+    });
+  }
+
+  get template() {
+    return editFormTemplate(this.point);
+  }
+}
