+++ webpack.config.js
@@ -38,6 +38,10 @@ module.exports = {
           },
         },
       },
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader']
+      },
     ]
   }
 };
